Refactor HashStoreFactory into HashStore module and update pytests
